pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["sambit <sambitkickers@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-pubsub = "^2.6.0"
loguru = "^0.5.3"
tweepy = "^3.10.0"
elasticsearch = "^7.13.3"
geojson_utils = "^0.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
